+++ library/src/test/java/com/twilio/video/util/ReflectionUtils.java
@@ -1,0 +1,16 @@
+package com.twilio.video.util;
+
+import java.lang.reflect.Field;
+import java.lang.reflect.Modifier;
+
+public class ReflectionUtils {
+    public static void setFinalStaticField(Field field, Object newValue) throws Exception {
+        field.setAccessible(true);
+
+        Field modifiersField = Field.class.getDeclaredField("modifiers");
+        modifiersField.setAccessible(true);
+        modifiersField.setInt(field, field.getModifiers() & ~Modifier.FINAL);
+
+        field.set(null, newValue);
+    }
+}